project-settings Ctrl+Z bug fixed
